requirements.txt
python-dotenv
pandas
numpy
yfinance
pyportfolioopt
openai
langchain
faiss-cpu
scipy
matplotlib
